+++ php-hdo/constants.php
@@ -1,0 +1,28 @@
+<?php
+
+
+ $openweather_api = "916e26607461b7d5b5c3fad8075fa22e";
+ $pwd="pa1e2";
+    
+    
+ //used in main, but let's keep it here   
+ $low_temp = 6;
+ $super_low_temp = -3;
+ $hi_temp = 28;
+ $super_hi_temp = 33;
+ $hi_wind = 10;
+ $super_hi_wind = 15;
+ $rain = 0.5;
+ $super_rain = 1.5;
+ $snow = 0.5;
+ $super_snow = 1.5;
+ 
+ $small_temp_diff = 2;
+ $big_temp_diff = 6;
+ 
+ $lat = "50.0973722";
+ $lon = "14.4074581";
+
+
+
+?>
