+++ php-hdo/utils1.php
@@ -151,6 +151,33 @@ function next_hdo_intervals($intervals)
   */
   return $hdo_intervals;
 }
-  
+ /////////////////////////////////////////
+ function cycling_index_check ($value, $direction, $limit, $hi_limit, $cycle_index)
+{
+
+    // the value does not exists, return cycle index as it is
+    if ($value != '')
+       {//value exists
+            
+      if ($direction == '>')
+      {
+          if ($value >= $limit)
+              $cycle_index = max ($cycle_index, 2);            
+          if ($value >= $hi_limit)
+              $cycle_index = max ($cycle_index, 3);
+      } //if vetsi
+              
+      if ($direction == '<')    
+      {
+          if ($value <= $limit)
+              $cycle_index = max ($cycle_index, 2);           
+          if ($value <= $hi_limit)
+              $cycle_index = max ($cycle_index, 3);
+      } //if vetsi
+         
+     } //value exists
+     
+     return $cycle_index;
+}  
 
 ?>
